Fix clashes between translator function in PythonModule.h/.cc
We have two global TU-local 'translator' functions in CMSSW:
One is in PythonModule.cc and one is in PythonModule.h.

While having TU-local functions in a header is already not recommended,
having them defined twice with the same name like this is causing
undefined behavior that breaks the linker (and also the rest
of clang in the future once someone includes the wrong file).

This patch is renaming the header function to at least fix the
current linking errors